+++ src/components/KMultiselect/KMultiselectItems.vue
@@ -2,7 +2,10 @@
   <KMultiselectItem
     v-for="item, idx in nonGroupedItems"
     :key="`${item.key ? item.key : idx}-item`"
+    ref="kMultiselectItem"
     :item="item"
+    @arrow-down="() => shiftFocus(item.key, 'down')"
+    @arrow-up="() => shiftFocus(item.key, 'up')"
     @selected="handleItemSelect"
   >
     <template #content>
@@ -12,6 +15,7 @@
       />
     </template>
   </KMultiselectItem>
+
   <div
     v-for="group in groups"
     :key="`${group}-group`"
@@ -22,8 +26,11 @@
     </span>
     <KMultiselectItem
       v-for="(item, idx) in getGroupItems(group)"
-      :key="`${item.key ? item.key : idx}-item`"
+      :key="`${item.key ? item.key : group + '-' + idx + '-item'}`"
+      ref="kMultiselectItem"
       :item="item"
+      @arrow-down="() => shiftFocus(item.key, 'down')"
+      @arrow-up="() => shiftFocus(item.key, 'up')"
       @selected="handleItemSelect"
     >
       <template #content>
@@ -38,13 +45,9 @@
 
 <script setup lang="ts">
 import type { PropType } from 'vue'
-import { computed } from 'vue'
+import { computed, ref } from 'vue'
 import KMultiselectItem from '@/components/KMultiselect/KMultiselectItem.vue'
 import type { MultiselectItem } from '@/types'
-
-interface MultiselectItemWithGroup extends MultiselectItem {
-  group: string
-}
 
 const props = defineProps({
   items: {
@@ -57,12 +60,50 @@ const props = defineProps({
 
 const emit = defineEmits(['selected'])
 
-const handleItemSelect = (item: MultiselectItem, isNew?: boolean) => emit('selected', item, isNew)
+const kMultiselectItem = ref<InstanceType<typeof KMultiselectItem>[] | null>(null)
+
+const handleItemSelect = (item: MultiselectItem) => emit('selected', item)
 
 const nonGroupedItems = computed((): MultiselectItem[] => props.items?.filter(item => !item.group))
-const groups = computed((): string[] => [...new Set((props.items?.filter(item => item.group) as unknown as MultiselectItemWithGroup[]).map(item => item.group))].sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase())))
+const groups = computed((): string[] => [...new Set((props.items?.filter(item => item.group))?.map(item => item.group!))].sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase())))
 
 const getGroupItems = (group: string) => props.items?.filter(item => item.group === group)
+
+const setFocus = (index: number = 0) => {
+  if (kMultiselectItem.value) {
+    if (!props.items[index].disabled) {
+      kMultiselectItem.value[index]?.$el?.querySelector('button').focus()
+    } else {
+      setFocus(index + 1)
+    }
+  }
+}
+
+const shiftFocus = (key: MultiselectItem['key'], direction: 'down' | 'up') => {
+  const index = props.items.findIndex(item => item.key === key)
+
+  if (index === -1) {
+    return // Exit if the item is not found
+  }
+
+  // determine step for navigation
+  const step = direction === 'down' ? 1 : -1
+  const isValidIndex = direction === 'down'
+    ? index + step < props.items.length
+    : index + step >= 0
+  if (isValidIndex) {
+    const nextIndex = index + step
+    if (props.items[nextIndex].disabled) {
+      // find the next valid index if the current one is disabled
+      shiftFocus(props.items[nextIndex].key!, direction)
+    } else {
+      // focus the button
+      kMultiselectItem.value?.[nextIndex]?.$el?.querySelector('button')?.focus()
+    }
+  }
+}
+
+defineExpose({ setFocus })
 </script>
 
 <style lang="scss" scoped>