+++ src/components/KSelect/KSelectItems.vue
@@ -48,7 +48,7 @@
 <script setup lang="ts">
 import { ref, computed } from 'vue'
 import type { PropType } from 'vue'
-import type { SelectItem, SelectItemWithGroup } from '@/types'
+import type { SelectItem } from '@/types'
 import KSelectItem from '@/components/KSelect/KSelectItem.vue'
 
 const props = defineProps({
@@ -68,9 +68,7 @@ const emit = defineEmits<{
 const handleItemSelect = (item: SelectItem) => emit('selected', item)
 
 const nonGroupedItems = computed((): SelectItem[] => props.items?.filter(item => !item.group))
-const groups = computed((): string[] =>
-  [...new Set((props.items?.filter(item => item.group) as unknown as SelectItemWithGroup[])
-    .map(item => item.group))].sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase())))
+const groups = computed((): string[] => [...new Set((props.items?.filter(item => item.group))?.map(item => item.group!))].sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase())))
 
 const getGroupItems = (group: string) => props.items?.filter(item => item.group === group)